+++ src/cli/main.rs
@@ -4,7 +4,7 @@
 use std::ffi::OsString;
 use std::sync::Arc;
 
-use crate::cli_shared::{cli::LogConfig, logger};
+use crate::cli_shared::logger;
 use crate::networks::ChainConfig;
 use crate::shim::address::{CurrentNetwork, Network};
 use crate::utils::io::ProgressBar;
@@ -30,7 +30,7 @@ where
         .block_on(async {
             match opts.to_config() {
                 Ok((mut config, _)) => {
-                    logger::setup_logger(&config.log, &opts);
+                    logger::setup_logger(&opts);
                     ProgressBar::set_progress_bars_visibility(config.client.show_progress_bars);
                     if opts.dry_run {
                         return Ok(());
@@ -70,7 +70,6 @@ where
                     }
                 }
                 Err(e) => {
-                    logger::setup_logger(&LogConfig::default(), &opts);
                     cli_error_and_die(format!("Error parsing config: {e}"), 1);
                 }
             }